+++ src/consumer/component/add-components/exceptions/index.js
@@ -6,6 +6,7 @@ import MissingComponentIdForImportedComponent from './missing-id-imported-compon
 import NoFiles from './no-files';
 import PathsNotExist from './paths-not-exist';
 import ExcludedMainFile from './excluded-main-file';
+import VersionShouldBeRemoved from './version-should-be-removed';
 
 export {
   DuplicateIds,
@@ -14,5 +15,6 @@ export {
   MissingComponentIdForImportedComponent,
   NoFiles,
   PathsNotExist,
+  VersionShouldBeRemoved,
   ExcludedMainFile
 };