+++ src/scope/exceptions/version-not-found.js
@@ -2,6 +2,7 @@
 import AbstractError from '../../error/abstract-error';
 
 export default class VersionNotFound extends AbstractError {
+  version: string;
   constructor(version: string) {
     super();
     this.version = version;